// src/clr-angular/utils/drag-and-drop/interfaces/drop-tolerance.interface.ts
/*
 * Copyright (c) 2016-2018 VMware, Inc. All Rights Reserved.
 * This software is released under MIT license.
 * The full license information can be found in LICENSE in the root directory of this project.
 */
export interface ClrDropToleranceInterface {
  top?: number;
  right?: number;
  bottom?: number;
  left?: number;
}